js backend: wrap generated in code in UMD-style factory
All generated code (including the runtime) is placed inside a UMD-style
factory method and the specified entry-points are exported.
This makes the usage of TeaVM generated JavaScript easier in bundlers.

The previous use-case of just loading it via a script-tag still works
fine and makes the exports available on the global window object like
before.